gateway/auth-service/requirements.txt
bcrypt>=4.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
PyJWT==2.8.0
pymongo==4.6.1
protobuf==4.25.1
